Reading double as double

+++ src/routines/ios.hpp
@@ -120,9 +120,9 @@ namespace nissa
   }
   
   //read a double with either a path or a file
-  template <class T> int master_fscan_double(T par)
+  template <class T> double master_fscan_double(T par)
   {
-    return master_fscan<int>(par,"%lg");
+    return master_fscan<double>(par,"%lg");
   }
 }
